Don't skip source generator if base has manual TestContext
This results in runtime errors that can be difficult to debug. Our error
message says to use `[UsesVerify]`, but the source generator doesn't
emit the property because it thinks it's already emitted.

Now instead this gives a compile time error with CS0108 (hides inherited
memeber) which explains what's going on pretty well.

Added unit test to capture behavior.